participants use sessionStorage
